python_code/requirements.txt
azure-identity>=1.12.0
azure-mgmt-resource>=23.0.0
azure-storage-blob>=12.14.0

# Optional (if you want to list subscriptions via SDK instead of supplying a subscription id):
# azure-mgmt-subscription>=0.6.0
